Refactor literals into their own submodule
