+++ src/lexer/literal/mod.rs
@@ -1,0 +1,4 @@
+pub mod literal;
+pub mod number;
+
+pub use literal::{lex_literal, Literal};
